refactor: action queue
