+++ TurboMinerVirtualTable.Unity/Assets/Scripts/Networking/ActionsQueue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b5f07d6c39711c45b4a3b0be5b725ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
