+++ TurboMinerVirtualTable.Unity/Assets/Scripts/Networking/Client.cs
@@ -1,15 +1,15 @@
 ﻿using Assets.Scripts.Networking;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Net.Sockets;
 using System.Threading;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 public class Client : MonoBehaviour
 {
+
+
     public string ClientName;
     public bool IsHost;
 
@@ -21,27 +21,16 @@ public class Client : MonoBehaviour
 
     private List<GameClient> players = new List<GameClient>();
     private DataSender dataSender;
+    private ActionsQueue actions;
 
     private Thread readThread;
     private bool threadStop;
-
-    private ConcurrentQueue<Action> mainThreadActions = new ConcurrentQueue<Action>();
 
     private void Start()
     {
         DontDestroyOnLoad(gameObject);
         dataSender = FindObjectOfType<DataSender>();
-    }
-
-    private void Update()
-    {
-        if (!mainThreadActions.IsEmpty)
-        {
-            while (mainThreadActions.TryDequeue(out var action))
-            {
-                action.Invoke();
-            }
-        }
+        actions = FindObjectOfType<ActionsQueue>();
     }
 
     public bool ConnectToServer(string host, int port)
@@ -114,133 +103,28 @@ public class Client : MonoBehaviour
     {
         Debug.Log($"Client: {data}");
 
-        var message = data.Split('|');
-
-        switch (message[0])
+        var messageFromServer = data.Split('|');
+        switch (messageFromServer[0])
         {
             case MessageCommands.Server.Who:
-                for (int i = 1; i < message.Length - 1; ++i)
+                for (int i = 1; i < messageFromServer.Length - 1; ++i)
                 {
-                    UserConnected(message[i], false);
+                    UserConnected(messageFromServer[i], false);
                 }
                 Send($"{MessageCommands.Client.Who}|{ClientName}|{(IsHost ? 1 : 0)}");
 
                 break;
 
             case MessageCommands.Server.Connected:
-                UserConnected(message[1], false);
+                UserConnected(messageFromServer[1], false);
                 if (IsHost)
                 {
                     dataSender.SendAllLobbySettings();
                 }
                 break;
-
-            case MessageCommands.Server.Start:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetTilesSettings(message[1]);
-                    MultiplayerManager.Instance.SetCorridorsSettings(message[2]);
-                    MultiplayerManager.Instance.SetMapSizeSettings(message[3], message[4]);
-                    MultiplayerManager.Instance.SetPlayerSettings(message[5]);
-                    SceneManager.LoadScene("Table");
-                });
-                break;
-
-            case MessageCommands.Server.WidthSettings:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetLobbyWidthDropdown(message[1]);
-                });
-                break;
-
-            case MessageCommands.Server.HeightSettings:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetLobbyHeightDropdown(message[1]);
-                });
-                break;
-
-            case MessageCommands.Server.TilesConfigName:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetLobbyTilesConfigDropdown(message[1]);
-                });
-                break;
-
-            case MessageCommands.Server.CorridorsConfigName:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetLobbyCorridorsConfigDropdown(message[1]);
-                });
-                break;
-
-            case MessageCommands.Server.LobbySettings:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetLobbyWidthDropdown(message[1]);
-                    MultiplayerManager.Instance.SetLobbyHeightDropdown(message[2]);
-                    MultiplayerManager.Instance.SetLobbyTilesConfigDropdown(message[3]);
-                    MultiplayerManager.Instance.SetLobbyCorridorsConfigDropdown(message[4]);
-                });
-                break;
-
-            case MessageCommands.Server.ElementPosition:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.SetElementsPositions(message[1]);
-                });
-                break;
-
-            case MessageCommands.Server.ElementStopDrag:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.StopElementDrag(int.Parse(message[1]));
-                });
-                break;
-
-            case MessageCommands.Server.ElementLayer:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.IncrementElementsLayers(message[1]);
-                });
-                break;
-
-            case MessageCommands.Server.ElementTurn:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.TurnElementOnOtherSide(int.Parse(message[1]));
-                });
-                break;
-
-            case MessageCommands.Server.ElementRotate:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.RotateElement(int.Parse(message[1]));
-                });
-                break;
-
-            case MessageCommands.Server.ElementDestroy:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.DestroyElement(int.Parse(message[1]));
-                });
-                break;
-
-            case MessageCommands.Server.RollDice:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.RollDice(int.Parse(message[1]));
-                });
-                break;
-
-            case MessageCommands.Server.StackRefill:
-                mainThreadActions.Enqueue(() =>
-                {
-                    MultiplayerManager.Instance.RefillStack(message[1]);
-                });
-                break;
         }
 
+        actions.AddActionAccordingTo(messageFromServer);
     }
 
     private void UserConnected(string name, bool host)
