+++ TurboMinerVirtualTable.Unity/Assets/Scripts/Ui/Lobby/LobbyDropdowns.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 369f980b8c8167742af322d010d363dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
